refactor(server): split team.ts — extract 4 star-player handlers (S27.8.23)
Second slice on `apps/server/src/routes/team.ts`. Extract the 4 Star
Player CRUD handlers into a dedicated module
`routes/team-star-player-handlers.ts` (488 lines).

Endpoints covered:
- `GET /:id/star-players` — handleListTeamStarPlayers
- `GET /:id/available-star-players` — handleListAvailableStarPlayers
  (budget + slots + pairs)
- `POST /:id/star-players` — handleHireStarPlayer (pair handling +
  match lock)
- `DELETE /:id/star-players/:starPlayerId` —
  handleDeleteTeamStarPlayer (removes pair if applicable)

All thematically cohesive (Star Player CRUD). Re-export from
`team.ts` to preserve the public API consumed by `team.test.ts`.

Leaf dependencies only (`prisma`, `sendError/sendSuccess`,
`updateTeamValues`, `getStarPlayerBySlug`/`DEFAULT_RULESET`/
`Ruleset` from `@bb/game-engine`, `validateStarPlayerHire`/
`getTeamAvailableStarPlayers`/`requiresPair` from
`../utils/star-player-validation`, `serverLog`, dynamic import
`getPlayerCost` from `team-value-calculator`). No cycle.

Cleanup of orphaned imports in `team.ts` (3 removed):
`validateStarPlayerHire`, `getTeamAvailableStarPlayers`,
`requiresPair`.

`team.ts` shrinks 2335 -> 1945 lines (-390 net lines, cumulative
-469 vs origin 2414 since S27.8.22 = -19%).

6 new smoke tests `team-star-player-handlers.test.ts`:
- exports x2 (named exports + re-export from team.ts)
- 404 gates: 4 handlers when team not found

- 1807/1807 server tests pass (was 1801, +6 new)
- typecheck clean
