+++ apps/server/src/routes/team.ts
@@ -23,12 +23,9 @@ import {
   getRerollCost,
 } from "@bb/game-engine";
 import {
-  validateStarPlayerHire,
   validateStarPlayerPairs,
   validateStarPlayersForTeam,
-  getTeamAvailableStarPlayers,
   calculateStarPlayersCost,
-  requiresPair,
 } from "../utils/star-player-validation";
 import { getRosterFromDb } from "../utils/roster-helpers";
 import { resolveRuleset } from "../utils/ruleset-helpers";
@@ -1598,426 +1595,39 @@ router.get("/:id/available-positions", authUser, handleListAvailablePositions);
 // STAR PLAYERS ENDPOINTS
 // =============================================================================
 
-// Endpoint pour obtenir les Star Players d'une equipe (S25.5q — ApiResponse<T>)
-export async function handleListTeamStarPlayers(
-  req: AuthenticatedRequest,
-  res: Response,
-): Promise<void> {
-  const teamId = req.params.id;
+// S27.8.23 — Handlers Star Players (list / list available / hire /
+// delete) extraits dans `routes/team-star-player-handlers.ts`.
+// Re-export pour preserver l'API publique consommee par les tests
+// d'integration (`team.test.ts`).
+export {
+  handleListTeamStarPlayers,
+  handleListAvailableStarPlayers,
+  handleHireStarPlayer,
+  handleDeleteTeamStarPlayer,
+} from './team-star-player-handlers';
+import {
+  handleListTeamStarPlayers as handleListTeamStarPlayersImpl,
+  handleListAvailableStarPlayers as handleListAvailableStarPlayersImpl,
+  handleHireStarPlayer as handleHireStarPlayerImpl,
+  handleDeleteTeamStarPlayer as handleDeleteTeamStarPlayerImpl,
+} from './team-star-player-handlers';
 
-  try {
-    const team = await prisma.team.findFirst({
-      where: { id: teamId, ownerId: req.user!.id },
-      include: { starPlayers: true },
-    });
-
-    if (!team) {
-      sendError(res, "Equipe introuvable", 404);
-      return;
-    }
-
-    const enrichedStarPlayers = team.starPlayers.map((sp: any) => {
-      const starPlayerData = getStarPlayerBySlug(sp.starPlayerSlug);
-      return {
-        id: sp.id,
-        slug: sp.starPlayerSlug,
-        cost: sp.cost,
-        hiredAt: sp.hiredAt,
-        ...starPlayerData,
-      };
-    });
-
-    sendSuccess(res, {
-      starPlayers: enrichedStarPlayers,
-      count: enrichedStarPlayers.length,
-    });
-  } catch (e: unknown) {
-    serverLog.error("Erreur lors de la recuperation des Star Players:", e);
-    sendError(res, "Erreur serveur", 500);
-  }
-}
-
-router.get("/:id/star-players", authUser, handleListTeamStarPlayers);
-
-// Endpoint pour obtenir les Star Players disponibles pour une equipe (S25.5aa — ApiResponse<T>)
-export async function handleListAvailableStarPlayers(
-  req: AuthenticatedRequest,
-  res: Response,
-): Promise<void> {
-  const teamId = req.params.id;
-
-  try {
-    const team = await prisma.team.findFirst({
-      where: { id: teamId, ownerId: req.user!.id },
-      include: { players: true, starPlayers: true },
-    });
-
-    if (!team) {
-      sendError(res, "Equipe introuvable", 404);
-      return;
-    }
-
-    const teamRuleset = (team.ruleset as Ruleset) ?? DEFAULT_RULESET;
-    const availableStarPlayers = getTeamAvailableStarPlayers(
-      team.roster,
-      teamRuleset,
-    );
-
-    const { getPlayerCost } = await import(
-      "../../../../packages/game-engine/src/utils/team-value-calculator"
-    );
-    const currentPlayersCost = team.players.reduce(
-      (total: number, player: any) => {
-        return total + getPlayerCost(player.position, team.roster, teamRuleset);
-      },
-      0,
-    );
-
-    const currentStarPlayersCost = team.starPlayers.reduce(
-      (total: number, sp: any) => {
-        return total + sp.cost;
-      },
-      0,
-    );
-
-    const budgetInPo = team.initialBudget * 1000;
-    const availableBudget =
-      budgetInPo - currentPlayersCost - currentStarPlayersCost;
-
-    const hiredSlugs = new Set(
-      team.starPlayers.map((sp: any) => sp.starPlayerSlug),
-    );
-    const totalPlayers = team.players.length + team.starPlayers.length;
-
-    const enrichedStarPlayers = availableStarPlayers.map((sp: any) => {
-      const isHired = hiredSlugs.has(sp.slug);
-      const canAfford = sp.cost <= availableBudget;
-      const hasRoomForOne = totalPlayers < 16;
-
-      const pairSlug = requiresPair(sp.slug);
-      let needsPair = false;
-      let pairStatus = null;
-
-      if (pairSlug) {
-        needsPair = true;
-        const pairHired = hiredSlugs.has(pairSlug);
-        const pairData = getStarPlayerBySlug(pairSlug, team.ruleset);
-        pairStatus = {
-          slug: pairSlug,
-          name: pairData?.displayName,
-          hired: pairHired,
-          cost: pairData?.cost || 0,
-        };
-      }
-
-      let canHire = !isHired && hasRoomForOne && canAfford;
-      if (needsPair && !pairStatus?.hired) {
-        const totalPairCost = sp.cost + (pairStatus?.cost || 0);
-        const hasRoomForPair = totalPlayers + 1 < 16;
-        canHire = !isHired && hasRoomForPair && totalPairCost <= availableBudget;
-      }
-
-      return {
-        ...sp,
-        isHired,
-        canHire,
-        needsPair,
-        pairStatus,
-      };
-    });
-
-    sendSuccess(res, {
-      availableStarPlayers: enrichedStarPlayers,
-      currentPlayerCount: team.players.length,
-      currentStarPlayerCount: team.starPlayers.length,
-      totalPlayers,
-      maxPlayers: 16,
-      availableBudget: Math.round(availableBudget / 1000),
-      totalBudget: team.initialBudget,
-    });
-  } catch (e: unknown) {
-    serverLog.error(
-      "Erreur lors de la recuperation des Star Players disponibles:",
-      e,
-    );
-    sendError(res, "Erreur serveur", 500);
-  }
-}
-
+router.get("/:id/star-players", authUser, handleListTeamStarPlayersImpl);
 router.get(
   "/:id/available-star-players",
   authUser,
-  handleListAvailableStarPlayers,
+  handleListAvailableStarPlayersImpl,
 );
-
-// Endpoint pour recruter un Star Player (S25.5ab — ApiResponse<T>)
-export async function handleHireStarPlayer(
-  req: AuthenticatedRequest,
-  res: Response,
-): Promise<void> {
-  const teamId = req.params.id;
-  const { starPlayerSlug } = req.body;
-
-  try {
-    const team = await prisma.team.findFirst({
-      where: { id: teamId, ownerId: req.user!.id },
-      include: { players: true, starPlayers: true },
-    });
-
-    if (!team) {
-      sendError(res, "Equipe introuvable", 404);
-      return;
-    }
-
-    const activeSelection = await prisma.teamSelection.findFirst({
-      where: {
-        teamId: teamId,
-        match: { status: { in: ["pending", "active"] } },
-      },
-    });
-
-    if (activeSelection) {
-      sendError(
-        res,
-        "Impossible de modifier cette equipe car elle est engagee dans un match en cours",
-        400,
-      );
-      return;
-    }
-
-    const { getPlayerCost } = await import(
-      "../../../../packages/game-engine/src/utils/team-value-calculator"
-    );
-    const teamRuleset = (team.ruleset as Ruleset) ?? DEFAULT_RULESET;
-    const currentPlayersCost = team.players.reduce(
-      (total: number, player: any) => {
-        return total + getPlayerCost(player.position, team.roster, teamRuleset);
-      },
-      0,
-    );
-
-    const currentStarPlayersCost = team.starPlayers.reduce(
-      (total: number, sp: any) => {
-        return total + sp.cost;
-      },
-      0,
-    );
-
-    const budgetInPo = team.initialBudget * 1000;
-    const availableBudget =
-      budgetInPo - currentPlayersCost - currentStarPlayersCost;
-
-    const validation = validateStarPlayerHire(
-      starPlayerSlug,
-      team.roster,
-      team.players.length,
-      team.starPlayers,
-      availableBudget,
-      teamRuleset,
-    );
-
-    if (!validation.valid) {
-      sendError(res, validation.error ?? "Validation echouee", 400);
-      return;
-    }
-
-    const starPlayer = validation.starPlayer!;
-
-    const pairSlug = requiresPair(starPlayerSlug);
-    const starPlayersToHire: Array<{ slug: string; cost: number }> = [
-      { slug: starPlayerSlug, cost: starPlayer.cost },
-    ];
-
-    if (pairSlug) {
-      const pairAlreadyHired = team.starPlayers.some(
-        (sp: any) => sp.starPlayerSlug === pairSlug,
-      );
-
-      if (!pairAlreadyHired) {
-        const pairData = getStarPlayerBySlug(pairSlug, team.ruleset);
-        if (!pairData) {
-          sendError(
-            res,
-            `Star Player partenaire '${pairSlug}' introuvable`,
-            400,
-          );
-          return;
-        }
-
-        const totalPlayers = team.players.length + team.starPlayers.length;
-        if (totalPlayers + 1 >= 16) {
-          sendError(
-            res,
-            "Pas assez de place pour recruter la paire (limite de 16 joueurs)",
-            400,
-          );
-          return;
-        }
-
-        const totalCost = starPlayer.cost + pairData.cost;
-        if (totalCost > availableBudget) {
-          sendError(
-            res,
-            `Budget insuffisant pour recruter la paire. Cout: ${(totalCost / 1000).toLocaleString()} K po, disponible: ${(availableBudget / 1000).toLocaleString()} K po`,
-            400,
-          );
-          return;
-        }
-
-        starPlayersToHire.push({ slug: pairSlug, cost: pairData.cost });
-      }
-    }
-
-    const createdStarPlayers = await Promise.all(
-      starPlayersToHire.map((sp) =>
-        prisma.teamStarPlayer.create({
-          data: {
-            teamId: teamId,
-            starPlayerSlug: sp.slug,
-            cost: sp.cost,
-          },
-        }),
-      ),
-    );
-
-    await updateTeamValues(prisma, teamId);
-
-    const updatedTeam = await prisma.team.findUnique({
-      where: { id: teamId },
-      include: { players: true, starPlayers: true },
-    });
-
-    const enrichedNewStarPlayers = createdStarPlayers.map((sp: any) => {
-      const starPlayerData = getStarPlayerBySlug(
-        sp.starPlayerSlug,
-        team.ruleset,
-      );
-      return {
-        id: sp.id,
-        slug: sp.starPlayerSlug,
-        cost: sp.cost,
-        hiredAt: sp.hiredAt,
-        ...starPlayerData,
-      };
-    });
-
-    sendSuccess(
-      res,
-      {
-        team: updatedTeam,
-        newStarPlayers: enrichedNewStarPlayers,
-        message:
-          enrichedNewStarPlayers.length > 1
-            ? `${enrichedNewStarPlayers
-                .map((sp: any) => sp.displayName)
-                .join(" et ")} recrutes avec succes`
-            : `${enrichedNewStarPlayers[0]?.displayName ?? starPlayerSlug} recrute avec succes`,
-      },
-      201,
-    );
-  } catch (e: any) {
-    serverLog.error("Erreur lors du recrutement du Star Player:", e);
-
-    if (e?.code === "P2002") {
-      sendError(res, "Ce Star Player est deja recrute dans cette equipe", 409);
-      return;
-    }
-
-    sendError(res, "Erreur serveur", 500);
-  }
-}
-
 router.post(
   "/:id/star-players",
   authUser,
   validate(addStarPlayerToTeamSchema),
-  handleHireStarPlayer,
+  handleHireStarPlayerImpl,
 );
-
-// Endpoint pour retirer un Star Player (S25.5v — ApiResponse<T>)
-export async function handleDeleteTeamStarPlayer(
-  req: AuthenticatedRequest,
-  res: Response,
-): Promise<void> {
-  const teamId = req.params.id;
-  const starPlayerId = req.params.starPlayerId;
-
-  try {
-    const team = await prisma.team.findFirst({
-      where: { id: teamId, ownerId: req.user!.id },
-      include: { players: true, starPlayers: true },
-    });
-
-    if (!team) {
-      sendError(res, "Equipe introuvable", 404);
-      return;
-    }
-
-    const activeSelection = await prisma.teamSelection.findFirst({
-      where: {
-        teamId: teamId,
-        match: { status: { in: ["pending", "active"] } },
-      },
-    });
-
-    if (activeSelection) {
-      sendError(
-        res,
-        "Impossible de modifier cette equipe car elle est engagee dans un match en cours",
-        400,
-      );
-      return;
-    }
-
-    const starPlayer = team.starPlayers.find(
-      (sp: any) => sp.id === starPlayerId,
-    );
-    if (!starPlayer) {
-      sendError(res, "Star Player introuvable", 404);
-      return;
-    }
-
-    const pairSlug = requiresPair(starPlayer.starPlayerSlug);
-    const starPlayersToRemove: string[] = [starPlayerId];
-
-    if (pairSlug) {
-      const pairStarPlayer = team.starPlayers.find(
-        (sp: any) => sp.starPlayerSlug === pairSlug,
-      );
-      if (pairStarPlayer) {
-        starPlayersToRemove.push(pairStarPlayer.id);
-      }
-    }
-
-    await prisma.teamStarPlayer.deleteMany({
-      where: { id: { in: starPlayersToRemove } },
-    });
-
-    await updateTeamValues(prisma, teamId);
-
-    const updatedTeam = await prisma.team.findUnique({
-      where: { id: teamId },
-      include: { players: true, starPlayers: true },
-    });
-
-    const removedCount = starPlayersToRemove.length;
-    sendSuccess(res, {
-      team: updatedTeam,
-      message:
-        removedCount > 1
-          ? `${removedCount} Star Players retires avec succes`
-          : "Star Player retire avec succes",
-    });
-  } catch (e: unknown) {
-    serverLog.error("Erreur lors du retrait du Star Player:", e);
-    sendError(res, "Erreur serveur", 500);
-  }
-}
-
 router.delete(
   "/:id/star-players/:starPlayerId",
   authUser,
-  handleDeleteTeamStarPlayer,
+  handleDeleteTeamStarPlayerImpl,
 );
 
 // Endpoint pour acheter avec la trésorerie (entre les matchs)
